added MainController to map access denied.., created Book/Review Services.., added AddBook, bookDetails, books, and denied.html

+++ PROG32758BookReviewsGroup20Assignment3/src/main/java/data/ReviewsRepository.java
@@ -3,5 +3,5 @@ package data;
 import models.Reviews;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReviewsRepository extends CrudRepository<Reviews, Integer> {
+public interface ReviewsRepository extends CrudRepository<Reviews, Long> {
 }
